requirements.txt
streamlit==1.32.2
langchain==0.1.13
langchain-google-genai==0.0.9
langchain-community==0.0.29
chromadb==0.4.24
datasets==2.18.0
pandas==2.2.1
sentence-transformers==2.5.1
tiktoken
python-dotenv
requests